+++ unity/EditorPlugin/AfterUnity56/EntryPoint.cs
@@ -11,12 +11,12 @@ namespace JetBrains.Rider.Unity.Editor.AfterUnity56
       if (UnityUtils.IsInBatchModeAndNotInRiderTests)
         return;
       
-      PluginEntryPoint.OnModelInitialization += UnitTesting.Initialization.OnModelInitializationHandler;
+      PluginEntryPoint.OnModelInitialization += Initialization.OnModelInitializationHandler;
       PluginEntryPoint.OnModelInitialization += Navigation.Initialization.OnModelInitializationHandler;
       PluginEntryPoint.OnModelInitialization += Packages.Initialization.OnModelInitializationHandler;
       AppDomain.CurrentDomain.DomainUnload += (EventHandler) ((_, __) =>
       {
-        PluginEntryPoint.OnModelInitialization -= UnitTesting.Initialization.OnModelInitializationHandler;
+        PluginEntryPoint.OnModelInitialization -= Initialization.OnModelInitializationHandler;
         PluginEntryPoint.OnModelInitialization -= Navigation.Initialization.OnModelInitializationHandler;
         PluginEntryPoint.OnModelInitialization -= Packages.Initialization.OnModelInitializationHandler;
       });
